requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-appconf==1.0.3
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-restframework==0.0.1
djangorestframework==3.9.3
Faker==1.0.4
gunicorn==19.9.0
idna==2.8
imagekit==1.0.6
login==0.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
pilkit==2.0
Pillow==6.0.0
psycopg2-binary==2.8.2
Pygments==2.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
text-unidecode==1.2
urllib3==1.24.1
whitenoise==4.1.2
